requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
crewai==0.86.0
crewai-tools==0.17.0
python-dotenv>=1.0.0
SQLAlchemy==2.0.36
neo4j>=5.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest-asyncio==0.23.3
pytest>=7.0.0
httpx>=0.24.0